Reorganized middleware a bit
